requirements.txt
nano-graphrag==0.4.8
pandas==2.0.3
numpy==1.24.3
networkx==3.1
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
pyyaml==6.0.1
tqdm==4.65.0
scikit-learn==1.3.0
openai==1.3.0
python-dotenv==1.0.0
datasets==2.14.0
transformers==4.33.0
torch==2.0.1
sentence-transformers==2.2.2